(#137) Move NciProfileInfo to module
* Moved NciProfileInfo into a js file in a module helpers folder
* Updated instances of window.NciProfileInfo to use the new exported functions

Closes #137
