--- site/src/js/library/modules/NciProfileInfo.js
@@ -1,3 +1,0 @@
-export function nciProfileInfo() {
-  // TBD move GetNCIProfileInfoUrl and GetNCIProfileInfoUrl in here and then export them as a module for access by other things
-}
